backend/requirements.txt
fastapi
uvicorn
openai
pydantic
python-dotenv
duckduckgo-search
